Refactor record function

+++ miningVN.py
@@ -41,43 +41,10 @@ def update_note(note_id, fields, tags=""):
     invoke("guiBrowse", query=f"nid:{note_id}")
 
 
-def record_audio(session, tags=""):
+def record(session, audio=False, screenshot=False, tags=""):
     curr_time = datetime.now()
     last_note = get_last_note()
-
-    note_time = datetime.fromtimestamp(last_note/1000)
-    diff = (curr_time - note_time).seconds
-    if diff > 2*60:
-        while True:
-            go_ahead = str(input("Last note was added over 2 minutes ago. Are you sure it is the one you want to update (y/n): ")).strip()
-            if go_ahead == "y":
-                break
-            elif go_ahead == "n":
-                return
-            else:
-                print("Invalid choice")
-
-    curr_time = curr_time.strftime('%Y-%m-%d_%H_%M_%S')
-
-    print("Starting recording in...")
-    print(2)
-    sleep(1)
-    print(1)
-    sleep(1)
-    print("Recording...")
-    try:
-        subprocess.run(shlex.split(f"sox -t coreaudio 'BlackHole 2ch' '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.mp3'"))
-    except KeyboardInterrupt:
-        pass
-
-    # last_note = get_last_note()
-    # if last_note > 0:
-    update_note(last_note, {"SentenceAudio": f"[sound:VN-{session}_{curr_time}.mp3]"}, tags)
-
-
-def screenshot(session, tags=""):
-    curr_time = datetime.now()
-    last_note = get_last_note()
+    update_fields = {}
 
     note_time = datetime.fromtimestamp(last_note/1000)
     diff = (curr_time - note_time).seconds
@@ -92,72 +59,50 @@ def screenshot(session, tags=""):
                 print("Invalid choice")
 
     curr_time = curr_time.strftime('%Y-%m-%d_%H_%M_%S')
-    try:
-        subprocess.run(shlex.split(f"screencapture -o -i -J window '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg'"))
-        subprocess.run(shlex.split(f"ffmpeg -i '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg' '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.webp'"))
-        subprocess.run(shlex.split(f"rm '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg'"))
-    except KeyboardInterrupt:
-        pass
 
-    # last_note = get_last_note()
-    # if last_note > 0:
-    update_note(last_note, {"Picture": f'<img alt="snapshot" src="VN-{session}_{curr_time}.webp">'}, tags)
+    if screenshot:
+        try:
+            subprocess.run(shlex.split(f"screencapture -o -i -J window '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg'"))
+            subprocess.run(shlex.split(f"ffmpeg -i '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg' '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.webp'"))
+            subprocess.run(shlex.split(f"rm '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg'"))
 
+            update_fields["Picture"] = f'<img alt="snapshot" src="VN-{session}_{curr_time}.webp">'
+        except KeyboardInterrupt:
+            pass
 
-def both(session, tags=""):
-    curr_time = datetime.now()
-    last_note = get_last_note()
+    if audio:
+        print("Starting recording in...")
+        for i in range(2, 0, -1):
+            print(i)
+            sleep(1)
+        print("Recording...")
 
-    note_time = datetime.fromtimestamp(last_note/1000)
-    diff = (curr_time - note_time).seconds
-    if diff > 5*60:
-        while True:
-            go_ahead = str(input("Last note was added over 5 minutes ago. Are you sure it is the one you want to update (y/n): ")).strip()
-            if go_ahead == "y":
-                break
-            elif go_ahead == "n":
-                return
-            else:
-                print("Invalid choice")
-
-    curr_time = curr_time.strftime('%Y-%m-%d_%H_%M_%S')
-    try:
-        subprocess.run(shlex.split(f"screencapture -o -i -J window '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg'"))
-        subprocess.run(shlex.split(f"ffmpeg -i '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg' '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.webp'"))
-        subprocess.run(shlex.split(f"rm '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.jpg'"))
-    except KeyboardInterrupt:
-        pass
-
-    print("Starting recording in...")
-    print(2)
-    sleep(1)
-    print(1)
-    sleep(1)
-    print("Recording...")
-
-    try:
-        subprocess.run(shlex.split(f"sox -t coreaudio 'BlackHole 2ch' '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.mp3'"))
-    except KeyboardInterrupt:
-        pass
+        try:
+            subprocess.run(shlex.split(f"sox -t coreaudio 'BlackHole 2ch' '/Users/ludo/Library/Application Support/Anki2/User 1/collection.media/VN-{session}_{curr_time}.mp3'"))
+        except KeyboardInterrupt:
+            update_fields["SentenceAudio"] = f"[sound:VN-{session}_{curr_time}.mp3]"
 
     # if last_note > 0:
-    update_note(last_note, {"SentenceAudio": f"[sound:VN-{session}_{curr_time}.mp3]", "Picture": f'<img alt="snapshot" src="VN-{session}_{curr_time}.webp">'}, tags)
+    if update_fields:
+        update_note(last_note, update_fields, tags)
 
 
 input("Make sure audio output device and terminal permission are set up correctly then press ENTER")
 
 session_name = input("Write session name: ")
 while True:
-    type_selection = str(input("[1] Screenshot\n[2] Audio\n[3] Both\nChoose: ")).strip()
+    type_selection = str(input("[1] Screenshot\n[2] Audio\n[3] Both\n[q] Quit\nChoose: ")).strip()
     tag = "VN "+session_name.replace(" ", "_")
     if type_selection[-1] == "N":
         tag += "NSFW "
         type_selection = type_selection[0]
     if type_selection == "2":
-        record_audio(session_name, tags=tag)
+        record(session_name, audio=True, tags=tag)
     elif type_selection == "1":
-        screenshot(session_name, tags=tag)
+        record(session_name, screenshot=True, tags=tag)
     elif type_selection == "3":
-        both(session_name, tags=tag)
+        record(session_name, audio=True, screenshot=True, tags=tag)
+    elif type_selection == "q":
+        break
     else:
         print("innvalid choice")
